[blocks] Move to standard QC components (progress & qc bars) in all upstream blocks
