+++ .changeset/famous-fishes-marry.md
@@ -1,0 +1,7 @@
+---
+'@milaboratories/milaboratories.ui-examples.ui': patch
+'@milaboratories/uikit': patch
+'@platforma-sdk/ui-vue': patch
+---
+
+[blocks] Move to standard QC components (progress & qc bars) in all upstream blocks
