+++ etc/blocks/ui-examples/ui/src/pages/AgGridVuePage/LinkComponent.ts
@@ -1,0 +1,10 @@
+import type { Component } from 'vue';
+import { h } from 'vue';
+
+export const LinkComponent: Component = {
+  props: ['params'],
+  setup(props) {
+    return () =>
+      h('a', { href: props.params.value, style: 'text-decoration: underline' }, props.params.value);
+  },
+};
