+++ etc/blocks/ui-examples/ui/src/pages/AgGridVuePage/index.ts
@@ -1,0 +1,1 @@
+export { default as AgGridVuePage } from './AgGridVuePage.vue';
