+++ sdk/ui-vue/src/AgGridVue/index.ts
@@ -1,1 +1,2 @@
 export * from './useAgGridOptionsSimple';
+export * from './createAgGridColDef';
